src/requirements.txt
django>=5.0
djangorestframework
djangorestframework-gis
psycopg2-binary
celery
django-celery-beat
flower
redis
django-tenants
django-storages
boto3
Pillow
django-environ
djangorestframework-simplejwt
gunicorn
whitenoise
django-cors-headers
reportlab
pytest-django
